// src/app/models/sjcs/CompensacionFacItem.ts
export class CompensacionFacItem {

    compensar: boolean;
    idPersona: string;
    numColegiado: string;
    nombre: string;
    idFactura: string;
    numeroFactura: string;
    fechaFactura: string;
    importeTotalFactura: string;
    importePendienteFactura: string;
    importeCompensado: string;
    importePagado: string;

    constructor() { }
}